Added Admob and Got Ads Working
I am having trouble getting the correct scoping to be able to inject the activity context into the admob class I made. The use cases are all set but I was unable to inject the dependencies properly. I need to move on but I commented out the code so I can look at it later when I have time.